socialbot/requirements.txt
openai
google-generativeai
chainlit
openai-agents
python-dotenv
fastapi

